fix(domain): 타입 변경

+++ src/main/java/subway/domain/RouteResultDTO.java
@@ -18,11 +18,11 @@ public class RouteResultDTO {
         return path;
     }
 
-    public Double getTimeWeight() {
-        return timeWeight;
+    public Integer getTimeWeight() {
+        return timeWeight.intValue();
     }
 
-    public Double getDistanceWeight() {
-        return DistanceWeight;
+    public Integer getDistanceWeight() {
+        return DistanceWeight.intValue();
     }
 }
